refactor: organize section folder by domain
Organized section components under project/ subdirectory:
- All section components are project-related, so grouped under project/

File moves:
- section/ProjectActionsSection.tsx → section/project/ProjectActionsSection.tsx
- section/ProjectLinkSection.tsx → section/project/ProjectLinkSection.tsx
- section/ProjectSummarySection.tsx → section/project/ProjectSummarySection.tsx
- section/ProjectTeamSection.tsx → section/project/ProjectTeamSection.tsx
- section/ProjectTechnologiesSection.tsx → section/project/ProjectTechnologiesSection.tsx

Updated imports in:
- src/app/components/sidebar/ProjectSidebar.tsx (all 5 section components)

This provides better organization and room for future portfolio or team sections.

Verified with pnpm lint and pnpm build.
